manifestmock: move container/pkg/commit mocks into helper
This commit moves the manifest mocks from "gen-manifests"
into a shared helper. This way we can reuse the same code
in `bootc-image-builder` so that we can compare the resulting
manifests that bib produces with the manifests that we will
generate with the "images" library when the manifest generation
moves from bib into images.