distro/rhel{9,10}: Azure 9.7+ & 10.1+: remove loglevel kernel option
We have a request to increase the kernel loglevel from 3 to 7 in the
current and future Azure marketplace images.  We decided to make this
change only in future releases (9.7+ and 10.1+) of the Azure marketplace
images.

The default RHEL kernel log level is 7, so dropping this option has the
desired effect, and is preferable to setting it explicitly (we prefer
relying on defaults whenever possible).

RHEL-102372