cloud/azure: facilitate mocking of individual azure clients
By moving from factories to clients directly mocking becomes easier, as
it removes the need to mock the factories as well.

By defining interfaces for each of the clients it becomes possible to
mock them. Only those parts of the azure apis that are actually used
should be part of the interfaces.