cloud/azure: mock the clients and add basic unit tests
This add basic unit tests for the golden path. Pollers and pagers are
mocked or configured to immediately be in their final state.

This is in preparation of adding boot capabilities for testing. The
steps to create a VM on azure are quite long, so having unit tests to
rely on are a must.