distro: fix bug in variable substitution for static distros
This fixes a bug that when a distro is not found via the
`match` mechanism but with an exact `name` match the
variables where not substituted. This caused an error in
image-builder-clis build.

We also need a reverse dependency test for ibcli in images but
that should be its own commit.