distro: add support for filesystem from default_fs_type
When we move to bootc we need a way to declare that certain
partitions use the distribution default filesystem as the
partition table itself is generic. This commit adds support
in the YAML to omit the filesystem type, i.e. to just write:
```yaml
    partition_table:
      test_arch:
        partitions:
          - payload_type: filesystem
            payload:
              mountpoint: "/"
```
which will then translated at YAML load time to the default
filesystem declared by the distribution. This allows us to
write a generic YAML based partition table in images that
is the same as the partition table in bootc-image-builder.