disk: pass default fs to FS customiaztions
Tradtionally we did not pass the filesystem with the
filesystem customizations which means that we always
created "xfs". This is problematic on bootc systems
that require us to use the container default filesystem
and arguably also on fedora that declares its default
fs to be "ext4" in our image definitions.

So this commit changes the behavior so that the defaultFs
is passed when creating custom mountpoints. This changes
nothing on rhel/centos but does change the behavior on
fedora - which should be fine as we can consider this a
bugfix.

The manifest diff shows that the fstype on fedora moves
from xfs to ext4 for the filesystem customizations.