linter: disable unnecessary field selector
Staticcheck seems to be enabled by default on CI and it has a quite
strict rule QF1008 that requires to avoid using embedded selector. Since
we like to be explicit quite often, let's disable this rule.

Example:

type A struct {
    FieldA string
}
type B struct {
    A
    FieldB string
}

// generates a linter issue:
a := B{"a", "b"}
x := a.A.FieldA