imagefilter: include repos in filter result
When returning the results of the image filter include the
repositories too. To re-create an image we need the distro
and the matching repositories, as they are currently only
loosely connected we need to tie them a bit more together
here. This will be needed by image-builder-cli for the
pkgsearch command.