distro: add `bootc` distro to build bootc images
This commit creates a new `pkg/distro/bootc` that registers into
the distrofacory and provides a "normal" distro.{Distro,Arch,ImgType}
abstraction. Because of the way we create distros (via the
reporegistry) it will need to be instanciated directly like this:
```
distro := bootc.NewBootcDistro(imgRef)
```
but once we work more on the distrofactory/reporegistry interaction
this will most likly change and the regular distrofactory can be
used.

A bootc based distro defines itself by inspecting the underlying
bootc image and sets various properties (like the default rootfs)
based on this container inspection.

In this version we just copy the code from `bootc-image-builder`
that is concerned with the manifest creation. This means there
is some duplication with the existing code in `images`. While
this is not ideal it makes it easier to do the followup cleanups
and it is easier to be sure that there are no behavioral changes.

In this initial version only the disk based image types are moving,
the installer type will follow. Unfortunately the git history of
the image-builder code is not included, I can't think of a good
way of doing it as we are only including parts of the file.

This is a first step towards https://github.com/osbuild/images/pull/1656