pkg/awscloud: don't return pointers from Register()
We used to return pointers from the `Register()` method for the AMI ID
and snapshot ID. The reasons were mostly convenience, because the AWS
SDK functions were also returning pointers to string. However, it is
more idiomatic to return plain value for simple types, so let's do it
in this case and adjust all other places accordingly.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>