pkg/awscloud: use platform.BootMode for Register() boot mode
Use platform.BootMode for Register() boot mode, instead of a string
value. platform.BootMode is the value that the images library uses
consistently internally, so passing it as is to the Register() method
should simplify a lot of things. The function will then take care of
converting the value to the appropriate EC2 boot mode value,
effectively abstracting this logic from the caller.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>