manifest: determine pipelines as payload/build in manifest
This commit makes it the job of the manifest package to
determine if a pipeline is a build pipeline or a payload
pipeline (instead of describing it in the image type and
the YAML as we do currently).

Moving the knownledge about payload/build pipelines out of
the imagetypes will make the information more precise, the
current way to report the payload/build is quite loose and
easy to get wrong.

We currently have a `manifest.Build` interface that is
implemented by the traditional `BuildrootFromPackages`
and the `BuildrootFromContainer` types. All payload
pipelines use them when doing:
```go
NewBase(name string, build Build)
```
and the type system will ensure that all build pipelines
used by payload pipelines implement the `Build`
interface.

With that we can implement "BuildPipelines()" by
just checking if a pipeline implements the "Build"
interface (and the inverse for PayloadPipelines).

What is nice is that this approach is exteremly simple
and does not require any new concepts (like the
PipelineRole in the previous PR#1732 that also attempts
to solve this problem). In the future we could do
more advanced things like building the dependency tree
of the pipelines to see what is build/payload (as
Tomáš suggested) but currently we don't need this as
all our build pipelines implement the build interface
and the payload ones don't.

I guess one downside might be if we ever build something
more complicated where a build pipeline would double as
a payload pipeline (i.e. something that implements both
interfaces) but we can think about this if/when we
do that (YAGNI and all that).