manifest: mark pipelines as payload/build in manifest package
This commit implements the suggestion from Achilleas in
https://github.com/osbuild/images/pull/1696#issuecomment-3156169592
and implements a "role" for each pipeline so that the image
types no longer need to know what pipeline is used for payload
and what for build.

Moving the knownledge about payload/build pipelines out of
the imagetypes will make the information more precise, the
current way to report the payload/build is quite loose and
easy to get wrong. By using the pipeline role we are more
precise here.