distro/rhel-{9,10}: include yum-utils in Azure CVM images
yum-utils is installed in pretty much all our cloud images.  We missed
adding it to CVM because we based the initial package list on
third-party kickstart files.  But now we should align them with our
other Azure images.

The yum-utils package was the first missing package to stand out in
testing.  Further alignment will be done at a later time.