many: rename additional modules to enabled
In the current codebase we have default states for all modules encoded
in a literal map in the `osbuild/anaconda_stage.go` file. This needs to
be configurable per distro or alternatively per image type.

This commit renames additional modules to enabled modules as in the
future this property will contain all modules to enable, not just the
additional ones.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>