osbuild/anaconda: drop defaultModuleStates
Remove the concept of defaultModuleStates. Instead the map is always
constructed solely from the enabled and disabled modules which now come
from the installer configuration.

Note, if you're reading this commit as part of the PR this is the change
that would start changing manifests; as we can see however no checksums
have changed here.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>