many: drop the hardcoded platforms
The hardcoded platforms like platform.AArch64 are not used anymore,
all platform information comes from YAML. So drop them from the
code and replace them in the tests with their dynamic equivalent.

This also removes the platform.X86_64 from a bunch of tests and
replaces that too. We *could* have left a compatibility X86
platform but given that we don't actually use it outside tests
and that everything comes from yaml it seems more dangerous
to keep than to remove as it might lead to expectation that
are just not true (i.e. platform.X86 is not used at all
anymore).