platform: rename PlatformConf -> platform.Data
The platform.PlatformConf is not a great name, platform.Data
is (a bit) better. Ideas welcome. Ideal would be
platform.Platform but that name is taken already by the
interface. We could consider droping the interface now that
we have only a single platform data struct but that is
probably something for a followup.