pkg/awscloud: make TerminateInstanceEC2() usable for osbuild-composer
Rename the AWS.TerminateInstanceEC2() -> AWS.TerminateInstancesEC2()
and make the method take a slice of instance IDs to terminate. The
timeout for waiting is now also configurable and if less than or equal
to zero, the method won't wait for instance termination. This will
enable the use of this method in osbuild-composer for service
maintenance purposes.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>