disk: remove local PartitioningMode type
Remove the partition.PartitioningMode type and import it from
osbuild/blueprint.

The partitioning mode is a blueprint option that can be used in
conjunction with Filesystem customizations to control how the partition
table will change to accommodate the new partitions (by default, a
partition table will be "upgraded" to LVM when a custom filesystem is
added, unless the partitioning mode is set to "raw").  Initially, it was
defined as an image option, and is still treated as an image option
here, while being a Blueprint option in all user-facing places
(osbuild-composer).  This commit does not remove it from the image
options, but drops the images copy of the enum type to use the
blueprint.PartitioningMode instead.