cloud/azure: support creating and destroying a VM
`CreateVM` creates each of the dependent resources and the VM
itself. The subnet of the VM is fixed, as a separate virtual network is
created for it. Adding ssh keys is supported.

This should cover the basic functionality needed to upload a vhd and
boot it in azure, which allows testing of all the azure images.

When uploading a vhd, only x86_64 is supported for now. To support
aarch64 the azure cloud package will need support for image galleries.