manifest: move more tests outside the base package
Move more tests outside the manifest package and into manifest_test.

The findStage() was previously defined in the manifest package and
exported through export_test.go to be used in the manifest_test package.
Now it is only defined in manifest_test.

The findStages() function has been moved from os_test.go to
manifest_test.go.  Currently it's only used in os_test.go but it fits in
the general file as a test utility function.

One major change in this commit is the exported serialization functions
in export_test.go.  Previously, we had exported Serialize() methods
defined, which we wrappers around the private serialize() method of
specific pipelines.  Now these have been replaced with functions that
take a Pipeline interface as first argument.  Two serialization
functions are defined: one that takes inputs (SerializeWith(), which
calls pipeline.serializeStart(inputs) and pipeline.serialize()) and one
that does not (Serialize(), which only calls pipeline.serialize()).