fedora/iot: bump size of raw images to 5 GiB
With the additional drivers and firmware included it seems the
iot-commit we build no longer fits onto a 4 GiB image. Let's up the
limit a bit.

Note that the iot-commit type that is built in image-builder is *not*
the same as the one produced by Fedora; it's generally only used in our
testing.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>