netinst: Drop anaconda module customization and add rootfs setting
This drops unused code setting the anaconda modules from the blueprint's
installer customization section -- the netinst does not allow any
anaconda module customization so that Anaconda's default behavior will
be enabled. ie. showing all the spokes in the UI

This also sets the rootfs type from the image definition's
installer_config, but that is normally overridden by the
image_config's iso_rootfs_type setting so it will act as a default when
image_config isn't set.