fedora: move x86_64 installer grub2
Move the condition that x86_64 installers use grub2 into the shared
installer config as it applies to all types.

Note that this *might* not apply to the simplified installer but it
seems that its pipeline doesn't do anything with it anyhow so it's safe
to share. If that behavior changes in the future; we'll deal with it
then?

No manifest changes due to this commit.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>