many: introduce `InstallerCustomizations`
Similar (or identical) to how we handle `OSCustomizations` in the `os`
pipeline generator this commit introduces `InstallerCustomizations`
which can be directly passed on instead of setting each option
separately.

Start with an arbitrary set of properties which are shared across all
images; expanded upon in follow-up commits.

When testing this commit I found that we never applied the additional
dracut module configuration to the live installer; instead of disabling
the application of this part of the installer config I've overridden it
in the imagetype definitions to make sure our artifacts stay the same.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>