distrodefs/rhel-10: drop kernel packages from the base azure package set
While this change affects all the Azure manifest checksums, it has no
effect on *resolved* manifests.  The 'kernel' package (or the one
specified in the 'default_kernel' in the image config) is added
automatically to images.

Generally, we never add the kernel to static image package sets.  This
allows for installing only one kernel when selecting one through the
blueprint kernel customization.