distrodefs/rhel-10: align Azure CVM package list with base Azure
With this commit, the following packages are added to the Azure CVM
image type (after depsolving):

- default-fonts-core-sans-4.1-3.el10.noarch.rpm
- fonts-filesystem-2.0.5-18.el10.noarch.rpm
- glibc-langpack-en-2.39-37.el10.x86_64.rpm
    (replaces glibc-minimal-langpack from the previous set)
- google-noto-fonts-common
- google-noto-sans-mono-vf-fonts
- google-noto-sans-vf-fonts
- google-noto-serif-vf-fonts
- hunspell
- hunspell-en
- hunspell-en-GB
- hunspell-en-US
- hunspell-filesystem
- langpacks-core-en
- langpacks-en
- langpacks-fonts-en
- nvme-cli
- redhat-mono-vf-fonts
- redhat-text-vf-fonts
- uuid

This does not completely align the CVM image type with the rest of the
Azure image types, but it adds the most important packages that were
identified to be missing.  Further refinement might be needed later.

HMS-9011