linter: disable error strings
Linter default configuration requires all error strings to start with
lower case letters. While this is a very useful pattern which makes
errors more suitable for wrapping, we currently have way too many places
where it is not the case so I am disabling the rule for now and let's
fix this later.