linter: skip possible nil pointer dereference
The linter does not recognize this pattern:

	if resp == nil {
		t.Skip("This test is for internal testing only")
	}

It misses a `return` statement and reporting false alerts on every use
of the `resp` variable. But the reality is that the t.Skip calls
t.SkipNow which causes the goroutine to immediately die, therefore the
rest of the function is never excecuted.