many: drop `ISORootKickstart`
Since a while we've had a boolean option that determines if the
kickstart file we write for certain image types ends up in the root of
an ISO or in the `/usr/share/anaconda/interactive-defaults.ks` path.

This changes things only for the `minimal-installer` in Fedora as all
other installer adjacent image types had this set to `true`.

I've tested the `minimal-installer` to still function.

Historically we introduced the `minimal-installer` (then called
`image-installer`) to test the Anaconda webui. At that point in time the
webui didn't support any kickstarts excepts the one in
`interactive-defaults.ks`.

Right now it only leads to confusion. I'll likely re-introduce the
ability to modify the kickstart path(s) through the installer configs in
the image definitions at a later point in time if we need it instead of
having a boolean toggle.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>