many: rename `workload` usage to `imgTypeCustomizations`
Now that the workload concept is removed and everything the
initial image type configuration is done via OSCustomization
we need to rename "workload" as it will be confusing without
the historic context (and arguably it was always a bit
confusing).

The new name needs to capture that the OSCustomizations
applied here come from the initial imagetype. I picked
`imgTypeCustomizations` as it seems to capture best
what this is doing. But its long and I'm open for more
ideas, I initially used `seed` as its short but its
probably too abstract and it captures the concept poorly.

Its also unclear how long-lived this name will be,
ideally it all will be merged into a single OSCustomizations
variable inside the various images but that needs to
happen in followups or the churn is too big.