test/scripts: new quickcheck script for CI image builds
Add a new test scripts that helps decide if any manifests will need to
be rebuilt when comparing two branches.  It determines this by checking
for diffs in paths that will, if changed, cause changes in the generated
manifests or builds configuration.

These checks are quicker than generating all real manifests and checking
against the build cache, but are only accurate because of our current
test setup.  If the test setup changes to depend on anything outside the
paths that are currently checked, the script will need to be updated.