distro,manifestgen: support custom depsolvers for distros
The bootc based distros need a custom depsolver because
they need to use the dnf from the container to ensure
that all the plugins/custom config is used.

So manifestgen now checks if the distro defines its own
way to depsolve and if so uses that instead of using
the default way.

Note that this means that bootc ISOs (that require rpms)
are incompatible with custom depsolve helpers.