Support payloads for raw devices
This adds a new Payload entity type "Raw", with a source pipeline and
path. If a partition contains such a payload, then the write-device
stage will be used to copy the named file onto the partition.

The goal is to use this from bootc-image-builder. If we find
that the target image uses android boot or ukiboot, then the aboot.img
file from the image will be written to the aboot_a/ukiboot_a partition.

This also extracts the shared code to generate the Devices option
for all the mkfs and the new write-device stages.