image,distro: drop `OSCustomizations` from (most) installers
This commit removes the `OSCustomizations` again from the the
live,net,ostree anaconda installers. They are not needed because
those images do not use any os customizations. Before we had the
`workload` everywhere but then a lot of the workload options
where not used in some image types (like the installers). With
this commit those are now removed.