many: kernel arguments options support for all installers
Support for kernel options in the iso got added to the tar
installer in osbuild#1727.

For consistency this commit also enables it for the other
installers via the pre-existing `AdditionalKernelOpts`
that is part of the `manifest.InstallerCustomizations`.

A helper to read the kernel options from the customizations
and default image type(s) is extracted as well.