image: use `AdditionalKernelOpts` in tar installer
We can drop the (now) special handling of kernel args via the
OSCustomizations in the tar installer. With the previous commit
we use the manifest.InstallerCustomizations.AdditionalKernelOpts
to handle the kernel arguments which is enough (this is also
validated by the fact that the existing TestTarInstallerKernelOpts
does not need to change).