image: add OSTreeSimplifiedInstaller.InstallerCustomimzations
The OSTreeSimplifiedInstaller is not an anaconda installer
but it still uses some of the anaconda structs. So lets
move it to use the same installer customizations.

Note that we cannot use the InstallerCustomizations.KernelAppend
with it as it will lead to a diff in the rhel-9.x
edge-simplified-installer image:
```yaml
--- /tmp/tmp07t_j3mj/ref/rhel_9.7-x86_64-edge_simplified_installer-edge_ostree_pull_device.json	2025-08-26 15:10:18.721496181 +0200
+++ /tmp/tmp07t_j3mj/new/rhel_9.7-x86_64-edge_simplified_installer-edge_ostree_pull_device.json	2025-08-26 15:10:21.201572008 +0200
@@ -1100,7 +1100,10 @@
                 "coreos.inst.isoroot=RHEL-9-7-0-BaseOS-x86_64",
                 "coreos.inst.install_dev=/dev/sda",
                 "coreos.inst.image_file=/run/media/iso/image.raw.xz",
-                "coreos.inst.insecure"
+                "coreos.inst.insecure",
+                "modprobe.blacklist=vc4",
+                "rw",
+                "coreos.no_persist_ip"
               ]
             },
             "isolabel": "RHEL-9-7-0-BaseOS-x86_64",
@@ -1259,7 +1262,10 @@
                 "coreos.inst.isoroot=RHEL-9-7-0-BaseOS-x86_64",
                 "coreos.inst.install_dev=/dev/sda",
                 "coreos.inst.image_file=/run/media/iso/image.raw.xz",
-                "coreos.inst.insecure"
+                "coreos.inst.insecure",
+                "modprobe.blacklist=vc4",
+                "rw",
+                "coreos.no_persist_ip"
               ]
             }
           }
```
This is because the edge-simplified-installer for rhel uses
the kernel options from the base image type only for the
deployed system but does not add it to the ISO kernel commandline.

Its unclear if we should unify this or not but this commit errs
on the side of caution.