image: fix kernel options compat in AnacondaOSTreeInstaller
The fedora `iot-installer` image sets default image type
kernel options that were previously not honored by the
image type. With the unification of the installer kernel
options it is no honored which leads to a manifest diff.

It is not clear if we want these kernel options or not
but this commit is conservative and keeps the old
behavior by just unsetting the kernel options in the
"iot-installer" image options.