image,manifest: pass InstallerCustomizations in NewAnaconda*
By passing the InstallerCustomiations into the various
image.NewAnaconda*() function we can simplify the code by
only passing a single struct (that is part of the anaconda
structs anyway).

This also means we can use the product/preview etc values
from the customizations and don't need to carry an extra
copy. So this is dropped too.