tools: add new `gen-bootc-diff` tool
This tool is meant as a temporary measure to ensure that we
produce identical manifests from the bootc integration of
images. We need it because we have no manifest checksums
for bootc based images (yet).

So the idea is:
1. have this tool validate that the manifests are identical
2. create manifests checks for bootc images (just like we
   do for regular images)
3. merge bib PR#1014 that switches disk generation to images
4. drop using the tool

Keep the tool in the backpocket because we will need it for
the ISO move to images too.