test: update manifest checksums
A bit of a rebase issue happened here.
PR #1750 updated the repository URLs (snapshots) for RHEL 8.4, modifying
all manifests.
PR #1773 added a new image type (netinst), which introduced new
manifests.

The netinst PR was merged a couple of days ago, adding the new
manifests to main, but #1750 was based on an old commit and didn't pick
those manifests up to change them.

We don't run the manifest checksum verification on the merge queue, nor
do we require it for merging to begin with, so this created the
inconsistency in main.

We knew this was a situation that could happen (see drawbacks at the end
of this issue description #1566), but I was pretty sure we wouldn't run
into it in reality.