fedora: do not mark images as firstboot
When we write `uninitialized` into `/etc/machine-id` it triggers
`systemd`'s first boot capabilities. This was wanted behavior at the
time but leads to interplay with other bits that are unexpected.

Specifically in the `minimal-raw` images we are using mount units and
have an `rw` root filesystem. This, in combination with the system
marked as booting for the first time triggers `systemd`'s first boot
full preset behavior [1].

This then means `systemd` overwrites all relevant configuration in
`/etc` with the presets. These presets do *not* include our mount units,
nor the `initial-setup` that this image normally uses.

An alternative is to drop mount units, write `/etc/fstab` and kernel
argument as `ro` again. This prevents `systemd` from performing its
reset (probably accidentally). However, in Fedora 43 the `setup` package
[2] was also changed so we now have a (correct) `!unprovisioned`
password hash for the `root` password. If this is present
`systemd-firstboot` will prompt for it; and we end up with duplicate
provisioning systems.

[1]: https://src.fedoraproject.org/rpms/systemd/blob/rawhide/f/systemd.spec#_839
[2]: https://src.fedoraproject.org/rpms/setup/c/7ced36d60b67c9e74f7951123225200597e3d2fa

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>