distro: ensure correct bootc image type Filename() output
This fixes a bug in the new bootc based imagetypes. The
filename was not returned with the correct extension.
The reason is that the the file extension is hardcoded in
pkg/image/bootc_disk.go, we have no way to access this
information the level when the image type is generated.

When the bootc imagetypes move into YAML the problem
hopefully solves itself as will then set the filename
in the YAML just like we do for the other image types
(this will need changes to bootc_disk.go then too).

So this commit needs to duplicate it for now. No
explicit test as this is indirectly tested via ibcli
and it will go away when moving into yaml based bootc
definitions (see https://github.com/osbuild/images/pull/1656
for an outline what that might look like).

(fwiw with https://github.com/osbuild/images/pull/1039
this wouldn't be an issue for ibcli).