github: give names to fedora versions in unit test matrix
Remove the matrix value from the names of tests and add the names
"previous" and "next" so we can keep the names of the actions stable.
This should save us from having to update the required tests in
repository settings every time we add or remove a Fedora version.

It seems there is no way to mark a whole step with a matrix as required.
We need to mark each individual step with its variables.

This commit also adds testing on Fedora branched (currently 43) and
rawhide (currently 44).  However, rawhide will not be marked as
required.  The job config is also marked with continue-on-error so it
doesn't fail the workflow run.