gen-manifests: add panic() handler
The current gen-manifest behavior is to not catch
panics. In part this is nice as it gives the full
backtrace. But it also lacks lots of content as
the panic happens without knowing for what image
type. So instead just handle a panic like an error
and how the image type that paniced and the error.