bib: add bibcontainer.Arch helper
We always have a local and running container with bib
so we can figure out the architecture for it. This is
useful for the bootc distro because with that we can
create the correct distro/arch/imagetypes combination.

Note that some images (like ubi9) do not show the
architecture when inspecting the image. So we need
a `podman exec uname -m` fallback for those.

Note also that it would be nice in the future to
inspect for multi arch images and actually populate
the bootc distro with the correct architectures. But
that is tricky as we currently introspect (run) each
container so we would have to download/run all the
distros just to do the inspection. This is generally
not possible (without qemu-user) and certainly not
practical.