distrodefs/rhel-7: add supported_blueprint_options to image types
Define the supported blueprint options for all the image types in
RHEL 7.  These are almost the same as the ones added to the disk image
types in Fedora, minus containers and openscap, which are generally not
supported in our RHEL 7 images.

The checkOptionsRhel7() function has also been updated to match the
Fedora version, but trimmed down to not include checks for ostree and
ISOs.