distrodefs/rhel-8: add {supported,required}_blueprint_options
Define the supported and required blueprint options for all the image
types in RHEL 8.  These are very similar to the ones added to the image
types in Fedora, with a few exceptions:
- Ignition is not supported for the edge-simplified-installer.
- RHEL distros define a "tar" image type that requires its own set of
  supported options.
- RHEL distros support customizations.rhsm.

The checkOptionsRhel8() function has also been updated to match the
Fedora and RHEL 7 version, with modifications for the RHEL 8 (e.g. swap
creation on aarch64).