distro/generic: fix ostree url requirement for Fedora IoT disk images
Fedora IoT disk images (iot-qcow2, iot-raw-xz) require an ostree URL to
pull a commit from, but we never checked for it.  When generating
manifests, the manifest instantiation function did produce an error, but
we should have been catching it in the options checker first.

Check for Bootable images as well as BootISO images with RPMOSTree when
validating the required ostree options.