distro/generic: remove image type name condition
The whole conditional block for the iot-simplified-installer was there
in the previous version of the function because it was also validating
the supported options for the image type.  Now that that validation is
covered elsewhere, we only need to check if the option combinations are
valid if and when the customizations (FDO or Ignition) are used.