github: use git status to verify checksum changes
When testing the head of the merge queue for manifest checksum changes,
use 'git status --porcelain' instead of 'git diff --exit-code'.  Using
diff does not exit with an error if there are untracked files in the
tree, which we want to catch.

The "Validate manifest checksums for every commit in the PR" will still
not catch file additions, because the rebase is not interrupted when
untracked files introduced in the tree.  To make sure we don't fail on
the merge queue when the PR is green, lets also run the "Validate
manifest checksums (HEAD)" check (which uses 'git status') on PRs.