bootc: add component tests based on fake containers
This commit introduces a fake bootc container that contains
just enough content to simulate a bootc image. It is based
on a "FROM scartch" container and does not require a network
connection. It will build a static go binary to provide the
"sleep" binary and the "bootc" binary that is expected by
the bootc introspection code in images.

The code to geneate fake bootc containers is put into
pkg/distro/bootc/bootctest
and can potentially be reused for manifest tests and
for ibcli/bib.

Thanks to Alex for suggesting the static go binary.