Support raw payloads in yaml partition info
Primarily, this removes the SourcePipeline option in the Raw payload
type, and instead always uses whatever is a "natural" source for the
type of image being built. This makes a lot more sense, because the
pipeline names is an internally generated thing and the user shouldn't
need to know what to use here.

We also add Raw handling to Partition.partitionTableFeatures() to avoid
this causing a panic.