internal: add randutil.String() helper
This commit adds a new `randutil.String()` helper
that will generate a random lower/upper/digest/symbols
string by default.

The implements is inspired by pythons random.choice(),
i.e. it is possible to pass a different set of strings
to generate the a way to generate the random string
from. This is useful to generate random container
IDs for example.