distrodefs/RHEL-10: use GCP guest tools for GCE image type (HMS-9236)
When we added initial support for GCE image types on RHEL-10, the GCP
guest tools build for el10 was not available. To workaround the
situation, we used cloud-init instead. The GCP guest tools for el10 have
been available since 27th of August, so let's use them for the el10 GCE
image type and bring it on par with the older RHEL releases.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>