Support setting what type of mount configuration to set in disk.yaml
This extends the current internal boolean mountUnits options to an
enum that also allows to disable both fstab and mount units.

Additionally, it also adds the option "mount_configuration" to the
bootc disk.yaml embedded config file, allowing bootc images to
configure this.

For now, distro.ImageConfig still has MountUnits, because changing
that would break the yaml format. However, this could be changed
in the future.

MOUNT_CONFIGURATION_FSTAB has value 0, to make the default match the
default for a mountUnit boolean (i.e. false).

Note: The previous MountUnits option also affected the generated
kernel config options in some cases. We do the same now even if the
option is set to none.