manifestgen: return manifest via Generator.Generate()
Simplify the way to use manifestgen to just return the
osbuild manifest bytes from `manifestgen.Generator.Generate()`
instead of the previous (indirect) way via `Options.Output`.

The previous way way was a bit too indirect and cumbersome
and this approach is more inline with what the existing
Serialize() is doing.

This will require updates in bib/ibcli.