bootctest: use Makefile to generate the fake bootc binary
This switches from the go code to build the fake bootc/sleep
binary to a Makefile. Its probably easier to read this way
(I say probably as Makefiles are not that big in the go world
but then the Makefile is a lot more concise).

Thanks to Alex for suggesting this.