github: make 'go mod tidy' part of the integration check
If 'go mod tidy' fails because of the osbuild/images update, we should
consider this part of the integration check and warn that the update
broke the package.  This is opposed to failing the job, which we want to
avoid when the cause is the changes made in the PR in images.