pkg/depsolvednf: decouple `dump` and `search` result from rpmmd
The Solver methods for fetching a repository metadata (`dump` command)
and for searching repository metadata (`search` command) were using
rpmmd.PackageList for unmarshalling responses from osbuild-depsolve-dnf.

Define an unexported type in `depsolvednf` package for the `dump` and
`search` command results and use it for unmarshalling in the respective
methods. This will allow us to redefine rpmmd.Package and
rpmmd.PackageList, without breaking the `depsolvednf` API
implementation.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>