os: Run dracut when config files are added to image
Previously the initrd was only recreated if fips mode was set, and
partitions were created. But it also should be run when dracut
configuration files are created using OSCustomizations.DracutConf.
In addition, these dracut config files need to be created before the
dracut stage runs.

With these changes the dracut config stages are created before
org.osbuild.dracut is run, and dracut is run when fips is enabled or
when configuration files have been added.

This changes quite a few manifests, fixing unnoticed issues with their
initrd not including changes as expected because dracut was either not
being run, or was running too soon.

eg. ec2 image should include sgdisk and blacklist the amdgpu and nouveau
modules, but this was not actually being applied.