reporegistry: add structured error ErrNoRepoFound
This commit adds a new reporegistry.ErrNoRepoFound that can
be used to find if a query for a repository failed because
its unavailable. This is required so that higher layers
can differenciate between failures like "invalid-arch/distro"
and repo-not-found.

It also updates the tests and adds an extra testcase for
unknown (vs invalid) distros.