osbuild: add Duration field to Status and parse it
Osbuild now emits duration of each stage/module in its status messages.
This patch adds Duration field to Status struct and parses it from
the JSON input. The field is of type time.Duration and is set to 0
if the duration is not present in the JSON (for example in older
osbuild versions).