distrodefs/rhel-9: add {supported,required}_blueprint_options
Define the supported and required blueprint options for all the image
types in RHEL 9.  These are very similar to the ones added to the image
types in RHEL 8, with a few exceptions:
- Ignition is supported for the edge-simplified-installer.

The checkOptionsRhel9() function has also been updated to match the new
functions for Fedora, RHEL 7, and RHEL 8, with appropriate
modifications.