distrodefs/rhel-10: add supported_blueprint_options
Define the supported and required blueprint options for all the image
types in RHEL 10.  These are very similar to the ones added to the image
types in RHEL 9.

The checkOptionsRhel10() function has also been updated to match the new
functions for Fedora, RHEL 7, RHEL 8, and RHEL 9 with appropriate
modifications (such as not checking for options associated with ostree
and edge image types).