docs: drop unused no_timer_check kernel parameter
This change removes the legacy no_timer_check kernel parameter from the
README to clean it up.

This parameter was needed in the past (circa 2014-2016) to work around issues
on older hardware, but it has not been necessary since the RHEL 7 era.

As explained in arch/x86/kernel/apic/io_apic.c, its original purpose was to
handle buggy SMP boards with incorrect mptable information about the timer IRQ.
On modern hypervisors like KVM, VMware, and Hyper-V, the kernel's timer
handling is robust, and the behavior already mimics the effect of this
parameter, making it redundant.

There is no reason to add this legacy parameter on RHEL 8 or later, and Fedora
19 and later. Removing it aligns with best practices by eliminating outdated
configuration, which helps prevent future confusion.